requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
spotipy==2.24.0
requests==2.32.3
Pillow==10.4.0
python-dotenv==1.0.1